+++ composeApp/src/commonMain/kotlin/io/music_assistant/client/ui/compose/home/players/PlayersPager.kt
@@ -1,6 +1,6 @@
 @file:OptIn(ExperimentalMaterial3Api::class)
 
-package io.music_assistant.client.ui.compose.home
+package io.music_assistant.client.ui.compose.home.players
 
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.core.tween
@@ -25,9 +25,6 @@ import androidx.compose.foundation.layout.wrapContentHeight
 import androidx.compose.foundation.layout.wrapContentSize
 import androidx.compose.foundation.pager.HorizontalPager
 import androidx.compose.foundation.pager.PagerState
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.VolumeMute
-import androidx.compose.material.icons.automirrored.filled.VolumeUp
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
@@ -55,9 +52,9 @@ import io.music_assistant.client.ui.compose.common.action.QueueAction
 import io.music_assistant.client.ui.compose.common.icons.VolumeIcon
 import io.music_assistant.client.ui.compose.common.icons.VolumeMutedIcon
 import io.music_assistant.client.ui.compose.common.rememberAnimatedDominantColor
-import io.music_assistant.client.ui.compose.home.players.GroupSettingsDialog
-import io.music_assistant.client.ui.compose.home.players.PlayerSelectionLayout
-import io.music_assistant.client.ui.compose.home.players.SelectPlayerDialog
+import io.music_assistant.client.ui.compose.home.CollapsibleQueue
+import io.music_assistant.client.ui.compose.home.HomeScreenViewModel
+import io.music_assistant.client.ui.compose.home.HorizontalPagerIndicator
 import io.music_assistant.client.utils.conditional
 
 @OptIn(ExperimentalMaterial3Api::class)
